+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hercules/cli
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hercules/core@0.2.2
+  - @hercules/tools@0.2.2
+  - @hercules/gateway@0.2.2
+
 ## 0.2.2
 
 ### Patch Changes