+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hercules/plugins
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hercules/core@0.2.2
+  - @hercules/tools@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes