+++ packages/scheduler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hercules/scheduler
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hercules/core@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes